shifted helpfulness to it's own eval function
